requirements.txt
absl-py==0.6.1
astor==0.7.1
certifi==2018.11.29
cycler==0.10.0
gast==0.2.0
grpcio==1.17.1
h5py==2.8.0
JumpWayMQTT==0.1.3
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
Markdown==3.0.1
matplotlib==3.0.2
numpy==1.15.4
opencv-python==3.4.4.19
paho-mqtt==1.3.1
pandas==0.23.4
protobuf==3.6.1
pyparsing==2.3.0
python-dateutil==2.7.5
pytz==2018.7
scikit-learn==0.20.1
scipy==1.1.0
six==1.12.0
sklearn==0.0
tensorboard==1.12.1
tensorflow-gpu==1.12.0
tensorflow-hub==0.2.0
termcolor==1.1.0
tqdm==4.28.1
Werkzeug==0.14.1
